# backend/app/api/routes_jobs.py
"""Analysis job endpoints.

POST /api/analyze                 multipart "file" -> {job_id}
GET  /api/jobs/{job_id}           status + current stage
GET  /api/jobs/{job_id}/result    full analysis JSON
"""

from __future__ import annotations

import asyncio
import re
import time
from typing import Literal
from urllib.parse import quote

from fastapi import APIRouter, Depends, File, Query, Response, UploadFile

from app.api.deps import get_service, rate_limited, require_api_key
from app.exporters.registry import get_format
from app.models.job import PROGRESS_STAGES, STAGE_LABELS, Job, JobStatus
from app.schemas.analysis import AnalysisResult
from app.schemas.jobs import (
    ErrorInfo,
    ErrorResponse,
    JobCreated,
    JobLinks,
    JobState,
    JobSummary,
    StageInfo,
)
from app.services.analysis_service import AnalysisService, JobFailedError
from app.services.result_view import transpose_result

router = APIRouter(prefix="/api", tags=["analysis"], dependencies=[Depends(require_api_key)])

WAIT_POLL_SECONDS = 0.5
_ERRORS: dict[int | str, dict[str, object]] = {
    code: {"model": ErrorResponse} for code in (401, 404, 409, 413, 415, 422, 429, 503)
}


def links(job_id: str) -> JobLinks:
    base = f"/api/jobs/{job_id}"
    return JobLinks(
        status=base, result=f"{base}/result", export_markdown=f"{base}/export?format=markdown"
    )


def stage_list(job: Job) -> list[StageInfo]:
    if job.status is JobStatus.COMPLETED:
        current = len(PROGRESS_STAGES)
    elif job.stage in PROGRESS_STAGES:
        current = PROGRESS_STAGES.index(job.stage)
    else:
        current = -1  # queued, or failed before a stage was recorded
    infos = []
    for i, stage in enumerate(PROGRESS_STAGES):
        state = "done" if i < current else "active" if i == current else "pending"
        if job.status is JobStatus.FAILED and state == "active":
            state = "failed"
        infos.append(StageInfo(id=stage, label=STAGE_LABELS[stage], state=state))
    return infos


def job_state(job: Job) -> JobState:
    error = (
        ErrorInfo(code=job.error_code or "analysis_failed", message=job.error_message or "")
        if job.status is JobStatus.FAILED
        else None
    )
    return JobState(
        job_id=job.id,
        status=job.status,
        stage=job.stage,
        stage_label=STAGE_LABELS[job.stage],
        stages=stage_list(job),
        filename=job.filename,
        file_size=job.file_size,
        created_at=job.created_at,
        updated_at=job.updated_at,
        summary=JobSummary(duration=job.duration, key=job.key, bpm=job.bpm),
        error=error,
        links=links(job.id),
    )


@router.post(
    "/analyze",
    status_code=202,
    response_model=JobCreated,
    responses=_ERRORS,
    dependencies=[Depends(rate_limited("analyze"))],
    summary="Upload an audio file for analysis",
)
async def analyze(
    response: Response,
    file: UploadFile = File(description="MP3, WAV, M4A, AAC, FLAC or OGG"),
    wait: int = Query(
        0,
        ge=0,
        le=300,
        description="Seconds to wait for completion; if it finishes in time the full result "
        "is included (handy for n8n / single-request integrations).",
    ),
    service: AnalysisService = Depends(get_service),
) -> JobCreated:
    submission = await service.submit(file, file.filename, file.content_type)
    job = submission.job
    deadline = time.monotonic() + wait
    while job.status in (JobStatus.QUEUED, JobStatus.PROCESSING) and time.monotonic() < deadline:
        await asyncio.sleep(WAIT_POLL_SECONDS)
        job = service.get_job(job.id)

    result = None
    if job.status is JobStatus.COMPLETED:
        response.status_code = 200
        if wait or submission.cached:
            result = service.get_result(job.id)
    elif job.status is JobStatus.FAILED and wait:
        raise JobFailedError(job.error_message, code=job.error_code)
    return JobCreated(
        job_id=job.id,
        status=job.status,
        cached=submission.cached,
        links=links(job.id),
        result=result,
    )


@router.get(
    "/jobs/{job_id}",
    response_model=JobState,
    responses=_ERRORS,
    dependencies=[Depends(rate_limited("read"))],
    summary="Job status and current processing stage",
)
async def get_job(job_id: str, service: AnalysisService = Depends(get_service)) -> JobState:
    return job_state(service.get_job(job_id))


@router.get(
    "/jobs/{job_id}/result",
    response_model=AnalysisResult,
    responses=_ERRORS,
    dependencies=[Depends(rate_limited("read"))],
    summary="Full analysis result, optionally transposed",
)
async def get_result(
    job_id: str,
    semitones: int = Query(0, ge=-11, le=11, description="Transpose by N semitones"),
    target_key: str | None = Query(
        None, max_length=24, description='Transpose to a key, e.g. "D" or "F# Minor"'
    ),
    service: AnalysisService = Depends(get_service),
) -> AnalysisResult:
    return transpose_result(service.get_result(job_id), semitones=semitones, target_key=target_key)


def _download_name(filename: str, key: str, semitones: int, mode: str, extension: str) -> str:
    stem = filename.rsplit(".", 1)[0] if "." in filename else filename
    parts = [stem]
    if semitones:
        parts.append(f"({key})")
    if mode == "beginner":
        parts.append("(beginner)")
    return " ".join(parts) + extension


@router.get(
    "/jobs/{job_id}/export",
    responses={**_ERRORS, 200: {"content": {"text/plain": {}, "text/markdown": {}}}},
    dependencies=[Depends(rate_limited("read"))],
    summary="Export the chord sheet (txt, markdown or json)",
)
async def export_result(
    job_id: str,
    export_format: str = Query("txt", alias="format", description="txt | markdown | json"),
    mode: Literal["original", "beginner"] = Query("original"),
    semitones: int = Query(0, ge=-11, le=11),
    target_key: str | None = Query(None, max_length=24),
    download: bool = Query(True, description="Send as attachment (false = inline text)"),
    service: AnalysisService = Depends(get_service),
) -> Response:
    export = get_format(export_format)
    result = transpose_result(
        service.get_result(job_id), semitones=semitones, target_key=target_key
    )
    result = result.model_copy(update={"view": result.view.model_copy(update={"chord_mode": mode})})
    body = export.render(result, mode)
    name = _download_name(
        result.metadata.filename, result.music.key, result.view.semitones, mode, export.extension
    )
    ascii_name = re.sub(r"[^A-Za-z0-9 ._()+-]", "_", name)
    disposition = "attachment" if download else "inline"
    headers = {
        "Content-Disposition": f'{disposition}; filename="{ascii_name}"; '
        f"filename*=UTF-8''{quote(name)}"
    }
    return Response(content=body, media_type=export.media_type, headers=headers)